refactor: ♻️ 원룸조회 API 코드 리팩토링 - 레포지토리 폴더삭제 - room서비스 폴더삭제
