+++ src/api/v1/room/room.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Req, UseGuards, Query } from '@nestjs/common';
-import { RoomService } from './service/room.service';
+import { RoomService } from './room.service';
 import { JwtAuthGuard } from '../auth/jwt/jwt.guard';
 import { Request } from 'express';
 import { RoomResponseDto } from './dto/RoomResponse.dto';
