+++ src/api/v1/room/room.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { RoomService } from './service/room.service';
+import { RoomService } from './room.service';
 import { RoomController } from './room.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { RoomRepository } from './repositroy/room.repository';
+import { RoomRepository } from './room.repository';
 import { Room } from '../../../libs/entity/room/room.entity';
 import { Favorite } from '../../../libs/entity/favorite/favorite.entity';
-import { RoomFilterService } from './service/roomFilter.service';
+import { RoomFilterService } from './roomFilter.service';
 
 @Module({
     imports: [TypeOrmModule.forFeature([Room, Favorite])],
